Rename: auth api 위치 변경 및 리팩토링
- auth api를 호출하는 위치가 useAuth -> useInitialze로 변경
- 이유: useAuth 에 위치하면 임포트하는 파일이 많아서 여러번 호출되는 오류가 있었음
- AUTH_REGEX, AUTH_INPUT_FIELD 파일 모듈로 분리
