+++ src/pages/UserCheckPassword.1.tsx
@@ -1,15 +1,12 @@
 import { SubmitHandler, useForm } from "react-hook-form";
 import { UserLoginStyle } from "./UserLogin";
 import Title from "@/components/user/Title";
-import InputField, { IInputField } from "@/components/inputField/InputField";
+import InputField from "@/components/inputField/InputField";
 import { IRessetPassword } from "@/models/auth.model";
 import SquareButton from "@/components/buttons/SquareButton";
-import { FaKey } from "react-icons/fa";
 
-const inputfield: Pick<IInputField, "icon" | "name"> = {
-  icon: <FaKey />,
-  name: "비밀번호 확인"
-};
+import { AUTH_INPUT_FIELD } from "@/utils/inputField";
+
 export const UserCheckPassword = () => {
   const {
     register,
@@ -25,7 +22,7 @@ export const UserCheckPassword = () => {
         <Title>비밀번호 확인</Title>
         <fieldset>
           <InputField
-            inputfield={inputfield}
+            inputfield={AUTH_INPUT_FIELD[2]}
             schema="auth"
             type="password"
             {...register("passwordCheck", { required: true })}
