+++ src/pages/UserLogin.tsx
@@ -1,25 +1,25 @@
-import InputField, { IInputField } from "@/components/inputField/InputField";
+import InputField from "@/components/inputField/InputField";
 import styled from "styled-components";
-import { FaSmile, FaKey } from "react-icons/fa";
 import Title from "@/components/user/Title";
 import SquareButton from "@/components/buttons/SquareButton";
 import { Link } from "react-router-dom";
 import { SubmitHandler, useForm } from "react-hook-form";
 import { ILogin } from "@/models/auth.model";
 import useAuth from "@/hooks/useAuth";
+import { AUTH_INPUT_FIELD } from "@/utils/inputField";
 
-const inputfield: IInputField[] = [
-  {
-    icon: <FaSmile />,
-    name: "이메일",
-    placeholder: "이메일을 입력해주세요"
-  },
-  {
-    icon: <FaKey />,
-    name: "비밀번호",
-    placeholder: "15자 이내로 입력해주세요"
-  }
-];
+// const inputfield: IInputField[] = [
+//   {
+//     icon: <FaSmile />,
+//     name: "이메일",
+//     placeholder: "이메일을 입력해주세요"
+//   },
+//   {
+//     icon: <FaKey />,
+//     name: "비밀번호",
+//     placeholder: "15자 이내로 입력해주세요"
+//   }
+// ];
 
 const UserLogin = () => {
   const {
@@ -41,7 +41,7 @@ const UserLogin = () => {
         <Title>로그인</Title>
         <fieldset>
           <InputField
-            inputfield={inputfield[0]}
+            inputfield={AUTH_INPUT_FIELD[0]}
             schema="auth"
             {...register("email", { required: true })}
           />
@@ -51,7 +51,7 @@ const UserLogin = () => {
         </fieldset>
         <fieldset>
           <InputField
-            inputfield={inputfield[1]}
+            inputfield={AUTH_INPUT_FIELD[1]}
             schema="auth"
             type="password"
             {...register("password", { required: true })}
